scripts/requirements.txt
# List your Python dependencies here
google-cloud-aiplatform
pyyaml